requirements.txt
Flask==3.0.3
gunicorn==22.0.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-flask==0.48b0
opentelemetry-instrumentation-requests==0.48b0
opentelemetry-exporter-otlp-proto-http==1.27.0
